🚚 Rename: apis.ts 파일 apis 폴더로 이동 및 auth로 파일명 수정
